Use a container's networks to find IP addresses
When adding a container, the container store would look for the top
level `IPAddress` field to find its IP. This works when the container is
using the default network, but fails when the container is using a
custom network. The container store now fetches the IPAddress field from
each network to determine the IPs associated with the container.